moved parseURL into extension helper
