+++ lib/extensions/queue-me.js
@@ -1,17 +1,17 @@
-  // var serviceHelper = require("./service-helper.js"),
 var extensionHelper = require("./extension-helper.js");
 
 var queueMeExtension = function(keyword, playQueue, request) {
     request = request || require("request");
     return {
         incoming: function(message, callback) {
-            var param = extensionHelper.CheckMessageForKeyword(message, keyword, true);
+            var unsanitisedParam = extensionHelper.CheckMessageForKeyword(message, keyword, true);
 
-            if (!param) {
+            if (!unsanitisedParam) {
                 callback(message);
                 return;
             }
-            var param = urlSanitiser(param);
+
+            var param = urlSanitiser(unsanitisedParam);
             var service = getService(param);
             var videoId = getVideoId(param);
             var personName = message.data.name;
@@ -54,56 +54,16 @@ function urlSanitiser(url){
 
 function getService(url){
     var youtubeURLs = ['youtu', 'youtube'];
-    parsedURL = parseURL(url);
-    if (youtubeURLs.indexOf(parsedURL.host) != -1){
+    var parsedURL = extensionHelper.parseURL(url);
+    if (youtubeURLs.indexOf(parsedURL.host) != -1) {
         return 'youtube';
-    };
+    }
 }
 
 function getVideoId(url){
     var regExp = /^.*((youtu.be\/)|(v\/)|(\/u\/\w\/)|(embed\/)|(watch\?))\??v?=?([^#\&\?]*).*/;
     var match = url.match(regExp);
     return match[7];
-}
-
-function parseURL(url){
-    var parsed_url = {};
-
-    if ( url === null || url.length === 0 ) {
-        return parsed_url;
-    }
-
-    var protocol_i = url.indexOf('://');
-    parsed_url.protocol = url.substr(0,protocol_i);
-
-    var remaining_url = url.substr(protocol_i + 3, url.length);
-    var domain_i = remaining_url.indexOf('/');
-    domain_i = domain_i == -1 ? remaining_url.length - 1 : domain_i;
-    parsed_url.domain = remaining_url.substr(0, domain_i);
-    parsed_url.path = domain_i == -1 || domain_i + 1 == remaining_url.length ? null : remaining_url.substr(domain_i + 1, remaining_url.length);
-
-    var domain_parts = parsed_url.domain.split('.');
-    switch ( domain_parts.length ){
-        case 2:
-            parsed_url.subdomain = null;
-            parsed_url.host = domain_parts[0];
-            parsed_url.tld = domain_parts[1];
-            break;
-        case 3:
-            parsed_url.subdomain = domain_parts[0];
-            parsed_url.host = domain_parts[1];
-            parsed_url.tld = domain_parts[2];
-            break;
-        case 4:
-            parsed_url.subdomain = domain_parts[0];
-            parsed_url.host = domain_parts[1];
-            parsed_url.tld = domain_parts[2] + '.' + domain_parts[3];
-            break;
-    }
-
-    parsed_url.parent_domain = parsed_url.host + '.' + parsed_url.tld;
-
-    return parsed_url;
 }
 
 function convert_time(duration) {
